updates for pr comments
