+++ server/common/default_config.py
@@ -171,8 +171,9 @@ dataset:
   user_annotations:
     enable: true
     type: local_file_csv
-    db_uri: postgresql://postgres:test_pw@localhost:5432
-    hosted_file_directory: s3://hosted-cellxgene-dev/user_annotations/
+    hosted_tiledb_array:
+        db_uri: null
+        hosted_file_directory: null
     local_file_csv:
       directory: null
       file: null
