+++ server/eb/app.py
@@ -74,8 +74,7 @@ def handle_config_from_secret(app_config):
     keyattrs = (
         ("flask_secret_key", "app__flask_secret_key"),
         ("oauth_client_secret", "authentication__params_oauth__client_secret"),
-        ("db_uri", "user_annotations__db_uri"),
-        ("hosted_file_directory", "user_annotations__hosted_file_directory")
+        ("db_uri", "user_annotations__hosted_tiledb_array__db_uri"),
     )
 
     for key, attr in keyattrs:
